move SocketIO.swift
